fixed issues and changed safetly limits to reflect hotter cyro
I fixed and issue with importing git temps when called from
fridge cycle. I fixed an issue for when you are trying to start
with a count that is not zero. I turned up the ADR voltage a little
so that it would actually reach the proper temperature. I changed
the safety temperature on teh 4K HTS to 10 K instead of 8.5