added new program to run fridge and ADR cycle in new way together
I added a new program ADR_4K_sink.py that runs the fridge cycle
while ramping the magnet. This way the heat of magnetization gets
sunk while the 4HE pump is hot to the 4K plate instead of sinking
directly to the 4He head. This is to extend the hold time of the
4He head. At the same time I added in a lot of fixes that were
long past due. This includes.
Simulataneuous plotting
Making the code more linear so that you can read it easier
Adding a PID loop (still needs to be tuned)
Fixing the Pause/Resume button